[강운덕] [LUCYSUS-1744] udp 포트 용도에 따라 분리함. span전용, 상태 정보용.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-tomcat-profiler/trunk@2323 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/test/java/com/nhn/pinpoint/profiler/AgentTest.java
@@ -27,7 +27,7 @@ public class AgentTest {
 //        config.readConfigFile();
         DefaultAgent agent = new DefaultAgent("", new DummyInstrumentation(), config);
 
-        short identifier = agent.getIdentifier();
+        int identifier = agent.getIdentifier();
         logger.info("{}", identifier);
     }
 }
